вроде заработало после инита

+++ urban-garbanzo/src/java/ru/garbanzo/urban/controller/MainServlet.java
@@ -23,6 +23,7 @@ import java.util.logging.Logger;
 import ru.garbanzo.urban.db.JDBCUtils;
 import ru.garbanzo.urban.edu.Answer;
 import ru.garbanzo.urban.edu.Question;
+import ru.garbanzo.urban.edu.Storage;
 import ru.garbanzo.urban.exception.JDBCException;
 import ru.garbanzo.urban.util.Utils;
 
@@ -81,7 +82,7 @@ public class MainServlet extends HttpServlet {
                 url = "/edit_question.jsp";
                 Utils.print("Servlet.edit_question", request.getParameterMap());
                 Utils.print(request.getParameter("qid"));
-                request.setAttribute("question", Question.getQuestionById(request.getParameter("qid")));
+                request.setAttribute("question", Question.getById(request.getParameter("qid")));
                 request.setAttribute("action", "update_question");
                 break;
             case "update_question":
@@ -99,14 +100,17 @@ public class MainServlet extends HttpServlet {
                 break;
 
             case "export":
-                Question.init(); // реинициализация, чтобы выгрузка была строго из БД
+                Storage.init(); // реинициализация, чтобы выгрузка была строго из БД
                 try {
+                    if (Storage.getJdbcException() != null) {
+                        throw Storage.getJdbcException();
+                    }
                     StringBuilder sb = new StringBuilder();
                         sb.append("DROP TABLE Question IF EXISTS;\r\n");
                         sb.append("CREATE TABLE Question (id INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY, realm VARCHAR(20), type int, text VARCHAR(2000));\r\n");
                         sb.append("\tDROP TABLE Answer IF EXISTS;\r\n");
                         sb.append("\tCREATE TABLE Answer (id INTEGER GENERATED ALWAYS AS IDENTITY PRIMARY KEY, questionId int, correct boolean, text VARCHAR(2000));\r\n");
-                    for (Question q: Question.getQuestionMap().values()) {
+                    for (Question q: Question.getMap().values()) {
                         Map<String, Object> state = q.getState();
                         sb.append("INSERT INTO Question (id");
                         for (String s: state.keySet()) {
